Add type registry, abstract some DWARF reading
We are adding support for proper types. The type system before had a
simple string representation; now we are adding a proper structured
representation. We add new functions for retrieving types in their structured form.
We are interested in types that show in exported function signatures. Our
approach is a two-pass: first we're going through the functions to add the
symbol name of the type, then we go through to resolve the structure of
each type. Backwards compatibility is maintained.