set up cargo workspace
the place we eventually want to get to is to have the dwarffi core library
be a standalone Rust crate that can be compiled into multiple different
front-ends. For example, one may want to generate python bindings, or
javascript bindings with ffi-napi, or javascript bindings with wasm-bindgen,
etc etc. There are lots of possible frontends, so we want to separate the
core logic (extracting function signatures and types from DWARF) from the
front-end logic, which may vary.

as part of this separation, dwarffi is a library-only crate. The executable
logic is moved to dwarffi-js.