Skip non-defining declarations
DWARF offers the following attribute:

A DW_AT_declaration attribute, whose value is a flag, which indicates whether
this entry represents a non-defining declaration of an object.

If we don't skip these, they appear as duplicates (potentially many times if
there are a lot of includes). Even though they are duplicates, they have
different offsets, so the registry will not deduplicate them (true for anything
that is declared: functions, struct definitions, etc.).