cleanup of primitive primitive types
primitive types had a ton of conditional logic for every permutation of
type names. in DWARF they are normalized so we didn't actually need that

we also removed the tests that hard-coded the count of functions in the
test C lib, since it kept failing whenever we needed to add new C code
to use in our tests.